Fixed descriptors, added overload for GetCampaigns
Fixed method descriptors for EmailCampaignService. Also added an overload for EmailCampaignService.GetCampaigns to continue accepting the same arguments as before.

Important: ALL instances of GetCampaigns now return a ResultSet object instead of an IList. To retrieve the actual list, you can fetch the Results member which is an IList. In addition, you can pass the ResultSet member Meta.Pagination back to ConstantContact.GetCampaigns as a single parameter to fetch the next page of results.